requirements.txt
# --- Core Data Science & Plotting ---
# Bibliothèques fondamentales pour la manipulation de données et la visualisation.
pandas>=2.1.0
numpy>=1.24.0,<2.0.0  
matplotlib>=3.7.0
seaborn>=0.13.0
PyYAML>=6.0

# --- Frameworks de Machine Learning & Deep Learning ---
scikit-learn>=1.3.0
joblib>=1.3.0
xgboost>=2.0.0
tensorflow>=2.15.0
torch>=2.1.0
transformers>=4.35.0

# --- Trading & Indicateurs Techniques ---
ccxt>=4.1.0
#pandas-ta>=0.3.14b0 
pandas-ta-openbb>=0.4.0

# --- Infrastructure & Connecteurs ---
redis>=5.0.0
websockets>=12.0
influxdb-client[ciso]>=1.40.0 
docker>=7.0.0 
aiohttp>=3.9.0

# --- API Web & Monitoring ---
fastapi[all]>=0.104.0 
pydantic>=2.5.0
prometheus-client>=0.19.0
praw>=7.7.0 

# --- Utilitaires ---
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0
watchfiles>=0.21.0